requirements.txt
Flask
Flask_Cors
python-dotenv
google-generativeai
Pillow